Consolidate approved-domains
